Reorganized tests & added isTrue/isFalse functions
